Improved partial hydration code (#199)
* Use foreach function

The previous function was very verbose. We can just get the list of
entries via a for each function

* moved await downwards

The await within the loop forces us to wait for the result of the
request. We only need that request once the component is loaded, so we
await the result there.

* Get element via target

No need to keep the id around in the object

* Rewrote loading code

Major change of the current logic. Functionally equivialent but more
compact by removing unnessesary variable / array creations.

* Fix: Applied eslint fixes

* Restore eager loading

* Skip lazy load code on eager only pages

On eager only pages the code will generate an constant 'if(false)' that
is never executed. terser as js minifier will remove the unreachable
code and inline the initComponent function.

* Also defer loading of props

Before (initially in 475223e) the props were loaded eagerly outside of
the intersection observer. I would suggest only loading them once
intersected, parallel to the component code so that it its fully lazy.

* Revert "Also defer loading of props"

This reverts commit ef765584e8e96c594e4a26f433a01cc9d9fad2be.

* Load props and components eagerly

Delays only prop decompression and component hydration via the IO

* Prefetch components by default, allow override

* Place props into json file

* Move component & prop fetching back into IO trigger